@@ -25,9 +25,9 @@
         }
     },
     "require": {
-        "php": "^7.3 || ~8.0.0 || ~8.1.0",
-        "laminas/laminas-stdlib": "^3.3",
-        "laminas/laminas-zendframework-bridge": "^1.0"
+        "php": "^7.4 || ~8.0.0 || ~8.1.0",
+        "ext-openssl": "*",
+        "laminas/laminas-stdlib": "^3.5"
     },
     "require-dev": {
         "laminas/laminas-coding-standard": "~2.3.0",
