@@ -41,7 +41,7 @@ class Rar extends AbstractCompressionAlgorithm
 
     /**
      * @param array $options (Optional) Options to set
-     * @throws Exception\ExtensionNotLoadedException if rar extension not loaded
+     * @throws Exception\ExtensionNotLoadedException If rar extension not loaded.
      */
     public function __construct($options = null)
     {
@@ -66,7 +66,7 @@ class Rar extends AbstractCompressionAlgorithm
      *
      * @param  string $callback
      * @return self
-     * @throws Exception\InvalidArgumentException if invalid callback provided
+     * @throws Exception\InvalidArgumentException If invalid callback provided.
      */
     public function setCallback($callback)
     {
@@ -139,7 +139,7 @@ class Rar extends AbstractCompressionAlgorithm
      *
      * @param  string $target
      * @return self
-     * @throws Exception\InvalidArgumentException if specified target directory does not exist
+     * @throws Exception\InvalidArgumentException If specified target directory does not exist.
      */
     public function setTarget($target)
     {
@@ -157,7 +157,7 @@ class Rar extends AbstractCompressionAlgorithm
      *
      * @param  string|array $content
      * @return string
-     * @throws Exception\RuntimeException if no callback available, or error during compression
+     * @throws Exception\RuntimeException If no callback available, or error during compression.
      */
     public function compress($content)
     {
@@ -182,8 +182,8 @@ class Rar extends AbstractCompressionAlgorithm
      *
      * @param  string $content
      * @return bool
-     * @throws Exception\RuntimeException if archive not found, cannot be opened,
-     *                                    or error during decompression
+     * @throws Exception\RuntimeException If archive not found, cannot be opened,
+     *                                    or error during decompression.
      */
     public function decompress($content)
     {
