@@ -10,7 +10,7 @@ interface FilterInterface
      * Returns the result of filtering $value
      *
      * @param  mixed $value
-     * @throws Exception\RuntimeException If filtering $value is impossible
+     * @throws Exception\RuntimeException If filtering $value is impossible.
      * @return mixed
      */
     public function filter($value);
